refactor(frontend): normalize editor utility naming
